+++ tests/testthat/test-utils.R
@@ -79,32 +79,30 @@ testthat::describe("add_facet_labels", {
 })
 
 testthat::describe("Module with decorators:", {
-  # We test it with tm_g_scatterplot as it requires decorators to be useful to users
-  create_scatterplot_module <- function(data, x_vars, y_vars, x_selected, y_selected, ...) {
-    tm_g_scatterplot(
-      x = list(
-        teal.transform::data_extract_spec(
-          dataname = "test_data",
-          select = teal.transform::select_spec(
-            choices = teal.transform::variable_choices(
-              data = isolate(data())[["test_data"]],
-              x_vars
-            ),
-            selected = x_selected,
-            multiple = FALSE
-          )
+  # We test it with tm_gtsummary as it requires decorators to be useful to users
+  create_gtsummary_module <- function(data, by_vars, include_vars, by_selected, include_selected, ...) {
+    tm_gtsummary(
+      by = teal.transform::data_extract_spec(
+        dataname = "test_data",
+        select = teal.transform::select_spec(
+          choices = teal.transform::variable_choices(
+            data = isolate(data())[["test_data"]],
+            by_vars
+          ),
+          selected = by_selected,
+          multiple = FALSE
         )
       ),
-      y = list(
+      include = list(
         teal.transform::data_extract_spec(
           dataname = "test_data",
           select = teal.transform::select_spec(
             choices = teal.transform::variable_choices(
               data = isolate(data())[["test_data"]],
-              y_vars
+              include_vars
             ),
-            selected = y_selected,
-            multiple = FALSE
+            selected = include_selected,
+            multiple = TRUE
           )
         )
       ),
@@ -114,12 +112,12 @@ testthat::describe("Module with decorators:", {
 
   it("one default decorator executes successfully", {
     data <- create_test_data(mtcars)
-    mod <- create_scatterplot_module(
+    mod <- create_gtsummary_module(
       data,
-      x_vars = c("mpg", "disp"),
-      y_vars = c("mpg", "disp"),
-      x_selected = "mpg",
-      y_selected = "disp",
+      by_vars = c("am", "gear"),
+      include_vars = c("carb", "cyl"),
+      by_selected = c("am"),
+      include_selected = c("carb", "cyl"),
       decorators = list(
         all = teal_transform_module()
       )
@@ -133,26 +131,11 @@ testthat::describe("Module with decorators:", {
       ),
       {
         session$setInputs(
-          "x-dataset_test_data_singleextract-select" = "mpg",
-          "y-dataset_test_data_singleextract-select" = "disp",
-          "log_x" = FALSE,
-          "log_y" = FALSE,
-          "rotate_xaxis_labels" = FALSE,
-          "ggtheme" = "gray",
-          "alpha" = 1,
-          "size" = 5,
-          "shape" = "circle",
-          "add_density" = FALSE,
-          "rug_plot" = FALSE,
-          "show_count" = FALSE,
-          "free_scales" = FALSE,
-          "pos" = 0.99,
-          "label_size" = 5,
-          "data_table_rows" = 10,
-          color = "#000000"
+          "by-dataset_test_data_singleextract-select" = "am",
+          "include-dataset_test_data_singleextract-select" = c("carb", "cyl")
         )
         session$flushReact()
-        testthat::expect_true(endsWith(get_code(session$returned()), "plot"))
+        testthat::expect_true(endsWith(get_code(session$returned()), "table"))
       }
     )
   })
@@ -160,12 +143,12 @@ testthat::describe("Module with decorators:", {
 
   it("one decorator failure is handled", {
     data <- create_test_data(mtcars)
-    mod <- create_scatterplot_module(
+    mod <- create_gtsummary_module(
       data,
-      x_vars = c("mpg", "disp"),
-      y_vars = c("mpg", "disp"),
-      x_selected = "mpg",
-      y_selected = "disp",
+      by_vars = c("am", "gear"),
+      include_vars = c("carb", "cyl"),
+      by_selected = c("am"),
+      include_selected = c("carb", "cyl"),
       decorators = list(
         all = teal_transform_module(server = function(id, data) {
           reactive({
@@ -183,23 +166,8 @@ testthat::describe("Module with decorators:", {
       ),
       {
         session$setInputs(
-          "x-dataset_test_data_singleextract-select" = "am",
-          "y-dataset_test_data_singleextract-select" = "gear",
-          "log_x" = FALSE,
-          "log_y" = FALSE,
-          "rotate_xaxis_labels" = FALSE,
-          "ggtheme" = "gray",
-          "alpha" = 1,
-          "size" = 5,
-          "shape" = "circle",
-          "add_density" = FALSE,
-          "rug_plot" = FALSE,
-          "show_count" = FALSE,
-          "free_scales" = FALSE,
-          "pos" = 0.99,
-          "label_size" = 5,
-          "data_table_rows" = 10,
-          color = "#000000"
+          "by-dataset_test_data_singleextract-select" = "am",
+          "include-dataset_test_data_singleextract-select" = c("carb", "cyl")
         )
         session$flushReact()
         testthat::expect_is(tryCatch(session$returned(), error = function(e) e), "shiny.silent.error")
@@ -209,13 +177,13 @@ testthat::describe("Module with decorators:", {
 
   it("Multiple decorators execute successfully", {
     data <- create_test_data(mtcars)
-    mod <- create_scatterplot_module(
+    mod <- create_gtsummary_module(
       data,
-      x_vars = c("am", "gear"),
-      y_vars = c("carb", "cyl"),
-      x_selected = "am",
-      y_selected = "carb",
-      decorators = list(plot = list(teal_transform_module(), teal_transform_module()))
+      by_vars = c("am", "gear"),
+      include_vars = c("carb", "cyl"),
+      by_selected = c("am"),
+      include_selected = c("carb", "cyl"),
+      decorators = list(table = list(teal_transform_module(), teal_transform_module()))
     )
 
     shiny::testServer(
@@ -226,41 +194,26 @@ testthat::describe("Module with decorators:", {
       ),
       {
         session$setInputs(
-          "x-dataset_test_data_singleextract-select" = "am",
-          "y-dataset_test_data_singleextract-select" = "gear",
-          "log_x" = FALSE,
-          "log_y" = FALSE,
-          "rotate_xaxis_labels" = FALSE,
-          "ggtheme" = "gray",
-          "alpha" = 1,
-          "size" = 5,
-          "shape" = "circle",
-          "add_density" = FALSE,
-          "rug_plot" = FALSE,
-          "show_count" = FALSE,
-          "free_scales" = FALSE,
-          "pos" = 0.99,
-          "label_size" = 5,
-          "data_table_rows" = 10,
-          color = "#000000"
+          "by-dataset_test_data_singleextract-select" = "am",
+          "include-dataset_test_data_singleextract-select" = c("carb", "cyl")
         )
         session$flushReact()
-        testthat::expect_true(endsWith(get_code(session$returned()), "plot"))
+        testthat::expect_true(endsWith(get_code(session$returned()), "table"))
       }
     )
   })
 
   it("Default and multiple decorators to one object execute successfully", {
     data <- create_test_data(mtcars)
-    mod <- create_scatterplot_module(
+    mod <- create_gtsummary_module(
       data,
-      x_vars = c("am", "gear"),
-      y_vars = c("carb", "cyl"),
-      x_selected = "am",
-      y_selected = "carb",
+      by_vars = c("am", "gear"),
+      include_vars = c("carb", "cyl"),
+      by_selected = c("am"),
+      include_selected = c("carb", "cyl"),
       decorators = list(
         all = teal_transform_module(),
-        plot = list(teal_transform_module(), teal_transform_module())
+        table = list(teal_transform_module(), teal_transform_module())
       )
     )
 
@@ -272,41 +225,25 @@ testthat::describe("Module with decorators:", {
       ),
       {
         session$setInputs(
-          "x-dataset_test_data_singleextract-select" = "am",
-          "y-dataset_test_data_singleextract-select" = "gear",
-          "log_x" = FALSE,
-          "log_y" = FALSE,
-          "rotate_xaxis_labels" = FALSE,
-          "ggtheme" = "gray",
-          "alpha" = 1,
-          "size" = 5,
-          "shape" = "circle",
-          "add_density" = FALSE,
-          "rug_plot" = FALSE,
-          "show_count" = FALSE,
-          "free_scales" = FALSE,
-          "pos" = 0.99,
-          "label_size" = 5,
-          "data_table_rows" = 10,
-          color = "#000000"
+          "by-dataset_test_data_singleextract-select" = "am",
+          "include-dataset_test_data_singleextract-select" = c("carb", "cyl")
         )
         session$flushReact()
-        testthat::expect_true(endsWith(get_code(session$returned()), "plot"))
+        testthat::expect_true(endsWith(get_code(session$returned()), "table"))
       }
     )
   })
-
   it("Default and one decorator to one object executes successfully", {
     data <- create_test_data(mtcars)
-    mod <- create_scatterplot_module(
+    mod <- create_gtsummary_module(
       data,
-      x_vars = c("am", "gear"),
-      y_vars = c("carb", "cyl"),
-      x_selected = "am",
-      y_selected = "carb",
+      by_vars = c("am", "gear"),
+      include_vars = c("carb", "cyl"),
+      by_selected = c("am"),
+      include_selected = c("carb", "cyl"),
       decorators = list(
         all = teal_transform_module(),
-        plot = teal_transform_module()
+        table = teal_transform_module()
       )
     )
 
@@ -318,26 +255,11 @@ testthat::describe("Module with decorators:", {
       ),
       {
         session$setInputs(
-          "x-dataset_test_data_singleextract-select" = "am",
-          "y-dataset_test_data_singleextract-select" = "gear",
-          "log_x" = FALSE,
-          "log_y" = FALSE,
-          "rotate_xaxis_labels" = FALSE,
-          "ggtheme" = "gray",
-          "alpha" = 1,
-          "size" = 5,
-          "shape" = "circle",
-          "add_density" = FALSE,
-          "rug_plot" = FALSE,
-          "show_count" = FALSE,
-          "free_scales" = FALSE,
-          "pos" = 0.99,
-          "label_size" = 5,
-          "data_table_rows" = 10,
-          color = "#000000"
+          "by-dataset_test_data_singleextract-select" = "am",
+          "include-dataset_test_data_singleextract-select" = c("carb", "cyl")
         )
         session$flushReact()
-        testthat::expect_true(endsWith(get_code(session$returned()), "plot"))
+        testthat::expect_true(endsWith(get_code(session$returned()), "table"))
       }
     )
   })